requirements.txt
asgiref==3.3.4
certifi==2021.5.30
chardet==3.0.4
click==8.0.1
colorama==0.4.4
decorator==4.4.2
dj-database-url==0.5.0
Django==3.2.1
django-cors-headers==3.7.0
django-debug-toolbar==3.2.1
django-restframework==0.0.1
djangorestframework==3.12.4
gensim==4.0.1
googletrans==4.0.0rc1
gunicorn==20.1.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.12.22
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
jellyfish==0.8.2
networkx==2.5.1
numpy==1.20.2
Pillow==8.2.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pytesseract==0.3.7
pytz==2021.1
regex==2021.4.4
requests==2.25.1
rfc3986==1.5.0
scipy==1.6.3
segtok==1.5.10
smart-open==5.0.0
sniffio==1.2.0
sqlparse==0.4.1
summarization==0.0.1
tabulate==0.8.9
tqdm==4.61.0
urllib3==1.26.5
whitenoise==5.2.0
yake==0.4.8
